+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # go
 
+## [1.2.0] - 21-08-2023
+- Feature: value.Converter can parse RFC3339 time format from string to time.Time
+- Improvement: add unit-test of interpretation "1" and "0" as bool values at value.Converter 
+
 ## [1.1.1] - 21-08-2023
-- Improvement: value.Converting use matching map for bool values
+- Improvement: value.Converter use matching map for bool values
 
 ## [1.1.0] - 25-06-2023
 - Init